Set up v1&v2 endpoints
