+++ eventmesh-runtime/src/main/java/org/apache/eventmesh/runtime/admin/handler/v1/TopicHandler.java
@@ -15,36 +15,37 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.runtime.admin.handler;
+package org.apache.eventmesh.runtime.admin.handler.v1;
 
+import org.apache.eventmesh.api.admin.TopicProperties;
 import org.apache.eventmesh.common.utils.JsonUtils;
+import org.apache.eventmesh.runtime.admin.handler.AbstractHttpHandler;
+import org.apache.eventmesh.runtime.admin.request.CreateTopicRequest;
+import org.apache.eventmesh.runtime.admin.request.DeleteTopicRequest;
 import org.apache.eventmesh.runtime.common.EventMeshHttpHandler;
 import org.apache.eventmesh.runtime.constants.EventMeshConstants;
 import org.apache.eventmesh.runtime.core.plugin.MQAdminWrapper;
+import org.apache.eventmesh.runtime.util.HttpResponseUtils;
 
-import java.net.URI;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
-import io.cloudevents.CloudEvent;
-import io.cloudevents.core.provider.EventFormatProvider;
-import io.cloudevents.jackson.JsonFormat;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.http.DefaultHttpHeaders;
 import io.netty.handler.codec.http.HttpHeaders;
 import io.netty.handler.codec.http.HttpRequest;
+import io.netty.handler.codec.http.HttpResponse;
+import io.netty.handler.codec.http.HttpResponseStatus;
 
 import lombok.extern.slf4j.Slf4j;
 
 /**
- * This class handles the {@code /event} endpoint, corresponding to the {@code eventmesh-dashboard} path {@code /event}.
+ * This class handles the {@code /topic} endpoint, corresponding to the {@code eventmesh-dashboard} path {@code /topic}, including the "Create Topic"
+ * and "Remove" buttons.
  * <p>
- * It is responsible for managing operations on events, including retrieving the event list and creating events.
- * <p>
- * The GET method supports querying events by {@code topicName}, and uses {@code offset} and {@code length} parameters for pagination.
+ * It provides functionality for managing topics, including retrieving the list of topics (GET), creating a new topic (POST), and deleting an existing
+ * topic (DELETE).
  * <p>
  * An instance of {@link MQAdminWrapper} is used to interact with the messaging system.
  *
@@ -53,8 +54,8 @@ import lombok.extern.slf4j.Slf4j;
  */
 
 @Slf4j
-@EventMeshHttpHandler(path = "/event")
-public class EventHandler extends AbstractHttpHandler {
+@EventMeshHttpHandler(path = "/topic")
+public class TopicHandler extends AbstractHttpHandler {
 
     private final MQAdminWrapper admin;
 
@@ -63,7 +64,7 @@ public class EventHandler extends AbstractHttpHandler {
      *
      * @param connectorPluginType The name of event storage connector plugin.
      */
-    public EventHandler(
+    public TopicHandler(
         String connectorPluginType) {
         super();
         admin = new MQAdminWrapper(connectorPluginType);
@@ -76,39 +77,34 @@ public class EventHandler extends AbstractHttpHandler {
 
     @Override
     protected void get(HttpRequest httpRequest, ChannelHandlerContext ctx) throws Exception {
-        HttpHeaders responseHeaders = new DefaultHttpHeaders();
-        responseHeaders.add(EventMeshConstants.HANDLER_ORIGIN, "*");
-        String queryString = URI.create(httpRequest.uri()).getQuery();
-        if (queryString == null || queryString.isEmpty()) {
-            writeUnauthorized(ctx, "");
-            return;
-        }
-        Map<String, String> queryMap = queryToMap(queryString);
-        String topicName = queryMap.get("topicName");
-        int offset = Integer.parseInt(queryMap.get("offset"));
-        int length = Integer.parseInt(queryMap.get("length"));
-        List<CloudEvent> eventList = admin.getEvent(topicName, offset, length);
-
-        List<String> eventJsonList = new ArrayList<>();
-        for (CloudEvent event : eventList) {
-            byte[] serializedEvent = Objects.requireNonNull(EventFormatProvider
-                    .getInstance()
-                    .resolveFormat(JsonFormat.CONTENT_TYPE))
-                .serialize(event);
-            eventJsonList.add(new String(serializedEvent, StandardCharsets.UTF_8));
-        }
+        List<TopicProperties> topicList = admin.getTopic();
+        String result = JsonUtils.toJSONString(topicList);
+        writeJson(ctx, result);
     }
 
     @Override
     protected void post(HttpRequest httpRequest, ChannelHandlerContext ctx) throws Exception {
+        Map<String, Object> body = parseHttpRequestBody(httpRequest);
+        Objects.requireNonNull(body, "body can not be null");
         HttpHeaders responseHeaders = new DefaultHttpHeaders();
+        responseHeaders.add(EventMeshConstants.CONTENT_TYPE, EventMeshConstants.APPLICATION_JSON);
         responseHeaders.add(EventMeshConstants.HANDLER_ORIGIN, "*");
-        String request = JsonUtils.toJSONString(parseHttpRequestBody(httpRequest));
-        byte[] rawRequest = request.getBytes(StandardCharsets.UTF_8);
-        CloudEvent event = Objects.requireNonNull(EventFormatProvider
-            .getInstance()
-            .resolveFormat(JsonFormat.CONTENT_TYPE)).deserialize(rawRequest);
-        admin.publish(event);
+        CreateTopicRequest createTopicRequest = JsonUtils.mapToObject(body, CreateTopicRequest.class);
+        String topicName = Objects.requireNonNull(createTopicRequest).getName();
+        admin.createTopic(topicName);
+        writeJson(ctx, "");
+    }
+
+    @Override
+    protected void delete(HttpRequest httpRequest, ChannelHandlerContext ctx) throws Exception {
+        Map<String, Object> body = parseHttpRequestBody(httpRequest);
+        Objects.requireNonNull(body, "body can not be null");
+        HttpHeaders responseHeaders = new DefaultHttpHeaders();
+        responseHeaders.add(EventMeshConstants.CONTENT_TYPE, EventMeshConstants.APPLICATION_JSON);
+        responseHeaders.add(EventMeshConstants.HANDLER_ORIGIN, "*");
+        DeleteTopicRequest deleteTopicRequest = JsonUtils.mapToObject(body, DeleteTopicRequest.class);
+        String topicName = Objects.requireNonNull(deleteTopicRequest).getName();
+        admin.deleteTopic(topicName);
         writeJson(ctx, "");
     }
 }
